rework so in future one middleware can be used for both ratelimiting and auth
