+++ apps/docs/libraries/ts/redwoodjs/middleware/ratelimiting.mdx
@@ -36,12 +36,14 @@ import type { withUnkeyOptions } from "@unkey/redwoodjs";
 
 export const registerMiddleware = () => {
   const options: withUnkeyOptions = {
-    ratelimitConfig: {
-      rootKey: process.env.UNKEY_ROOT_KEY,
-      namespace: "my-app",
-      limit: 1,
-      duration: "30s",
-      async: true,
+    ratelimit: {
+      config: {
+        rootKey: process.env.UNKEY_ROOT_KEY,
+        namespace: "my-app",
+        limit: 1,
+        duration: "30s",
+        async: true,
+      },
     },
   };
 
@@ -61,12 +63,14 @@ import type { withUnkeyOptions } from "@unkey/redwoodjs";
 
 export const registerMiddleware = () => {
   const options: withUnkeyOptions = {
-    ratelimitConfig: {
-      rootKey: process.env.UNKEY_ROOT_KEY,
-      namespace: "my-app",
-      limit: 1,
-      duration: "30s",
-      async: true,
+    ratelimit: {
+      config: {
+        rootKey: process.env.UNKEY_ROOT_KEY,
+        namespace: "my-app",
+        limit: 1,
+        duration: "30s",
+        async: true,
+      },
     },
   };
 
@@ -128,14 +132,16 @@ export const supabaseRatelimitIdentifier = (req: MiddlewareRequest) => {
 
 export const registerMiddleware = () => {
   const options: withUnkeyOptions = {
-    ratelimitConfig: {
-      rootKey: process.env.UNKEY_ROOT_KEY,
-      namespace: "my-app",
-      limit: 1,
-      duration: "30s",
-      async: true,
+    ratelimit: {
+      config: {
+        rootKey: process.env.UNKEY_ROOT_KEY,
+        namespace: "my-app",
+        limit: 1,
+        duration: "30s",
+        async: true,
+      },
+      ratelimitIdentifierFn: supabaseRatelimitIdentifier,
     },
-    ratelimitIdentifierFn: supabaseRatelimitIdentifier,
   };
   const unkeyMiddleware = withUnkey(options);
   const supabaseAuthMiddleware = createSupabaseAuthMiddleware({
@@ -154,17 +160,19 @@ export const registerMiddleware = () => {
 ```ts
 export const registerMiddleware = () => {
   const options: withUnkeyOptions = {
-    ratelimitConfig: {
-      rootKey: process.env.UNKEY_ROOT_KEY,
-      namespace: "my-app",
-      limit: 1,
-      duration: "30s",
-      async: true,
-    },
-    ratelimitExceededResponseFn: (_req: MiddlewareRequest) => {
-      return new MiddlewareResponse("Custom Rate limit exceeded message", {
-        status: 429,
-      });
+    ratelimit: {
+      config: {
+        rootKey: process.env.UNKEY_ROOT_KEY,
+        namespace: "my-app",
+        limit: 1,
+        duration: "30s",
+        async: true,
+      },
+      ratelimitExceededResponseFn: (_req: MiddlewareRequest) => {
+        return new MiddlewareResponse("Custom Rate limit exceeded message", {
+          status: 429,
+        });
+      },
     },
   };
   const unkeyMiddleware = withUnkey(options);
@@ -178,14 +186,14 @@ export const registerMiddleware = () => {
 ```ts
 export const registerMiddleware = () => {
   const options: withUnkeyOptions = {
-    ratelimitConfig: {
+    ratelimit: {config: {
       rootKey: process.env.UNKEY_ROOT_KEY,
       namespace: "my-app",
       limit: 1,
       duration: "30s",
       async: true,
     },
-    matcher: ["/blog-post/:slug(\\d{1,})"],
+    matcher: ,
     ratelimitErrorResponseFn: (_req: MiddlewareRequest) => {
       return new MiddlewareResponse("Custom Error message when rate limiting", {
         status: 500,
@@ -194,6 +202,6 @@ export const registerMiddleware = () => {
   };
   const unkeyMiddleware = withUnkey(options);
 
-  return [unkeyMiddleware];
+  return [[unkeyMiddleware, "/blog-post/:slug(\\d{1,})"]];
 };
 ```
