+++ packages/redwoodjs/src/ratelimit/__tests__/withUnkey.test.ts
@@ -4,7 +4,7 @@ import { Ratelimit } from "@unkey/ratelimit";
 import type { RatelimitConfig, RatelimitResponse } from "@unkey/ratelimit";
 import { assert, describe, expect, it, vi } from "vitest";
 import withUnkey from "../index";
-import type { withUnkeyOptions } from "../index";
+import type { withUnkeyOptions } from "../types";
 
 /**
  * Mock the Ratelimit class
@@ -48,9 +48,11 @@ const defaultRateLimitConfig: Pick<RatelimitConfig, "rootKey" | "limit" | "durat
 describe("withUnkey", () => {
   it("should not rate limit", async () => {
     const options: withUnkeyOptions = {
-      ratelimitConfig: {
-        namespace: "192.168.1.1",
-        ...defaultRateLimitConfig,
+      ratelimit: {
+        config: {
+          namespace: "192.168.1.1",
+          ...defaultRateLimitConfig,
+        },
       },
     };
     const unkeyMiddleware = withUnkey(options);
@@ -67,9 +69,11 @@ describe("withUnkey", () => {
     // it does not match the identifier generated by the withUnkey middleware
     // which is a defaulted value
     const options: withUnkeyOptions = {
-      ratelimitConfig: {
-        namespace: "my-app",
-        ...defaultRateLimitConfig,
+      ratelimit: {
+        config: {
+          namespace: "my-app",
+          ...defaultRateLimitConfig,
+        },
       },
     };
     const unkeyMiddleware = withUnkey(options);
@@ -85,13 +89,15 @@ describe("withUnkey", () => {
     // not to be exceeded (status 200) because it
     // matches the identifies generated by the withUnkey middleware
     const options: withUnkeyOptions = {
-      ratelimitConfig: {
-        namespace: "abcdefg",
-        ...defaultRateLimitConfig,
-      },
+      ratelimit: {
+        config: {
+          namespace: "abcdefg",
+          ...defaultRateLimitConfig,
+        },
 
-      ratelimitIdentifierFn: (_req: MiddlewareRequest) => {
-        return "abcdefg"; // matches namespace so no rate limit for this test
+        ratelimitIdentifierFn: (_req: MiddlewareRequest) => {
+          return "abcdefg"; // matches namespace so no rate limit for this test
+        },
       },
     };
     const unkeyMiddleware = withUnkey(options);
@@ -106,13 +112,15 @@ describe("withUnkey", () => {
     // here the namespace will cause the rate limit to be exceeded (status 429)
     // because it does not match the identifier generated by the withUnkey middleware
     const options: withUnkeyOptions = {
-      ratelimitConfig: {
-        namespace: "12345",
-        ...defaultRateLimitConfig,
-      },
+      ratelimit: {
+        config: {
+          namespace: "12345",
+          ...defaultRateLimitConfig,
+        },
 
-      ratelimitIdentifierFn: (_req: MiddlewareRequest) => {
-        return "abcdefg";
+        ratelimitIdentifierFn: (_req: MiddlewareRequest) => {
+          return "abcdefg";
+        },
       },
     };
     const unkeyMiddleware = withUnkey(options);
@@ -127,15 +135,17 @@ describe("withUnkey", () => {
     // here the namespace will cause the rate limit to be exceeded (status 429)
     // because it does not match the identifier generated by the withUnkey middleware
     const options: withUnkeyOptions = {
-      ratelimitConfig: {
-        namespace: "exceeded",
-        ...defaultRateLimitConfig,
-      },
+      ratelimit: {
+        config: {
+          namespace: "exceeded",
+          ...defaultRateLimitConfig,
+        },
 
-      ratelimitExceededResponseFn: (_req: MiddlewareRequest) => {
-        return new MiddlewareResponse("Custom Rate limit exceeded message", {
-          status: 429,
-        });
+        ratelimitExceededResponseFn: (_req: MiddlewareRequest) => {
+          return new MiddlewareResponse("Custom Rate limit exceeded message", {
+            status: 429,
+          });
+        },
       },
     };
     const unkeyMiddleware = withUnkey(options);
@@ -151,13 +161,15 @@ describe("withUnkey", () => {
     // here the the identifier function will throw an error
     // which will cause a teh default error response to be returned
     const options: withUnkeyOptions = {
-      ratelimitConfig: {
-        namespace: "exceeded",
-        ...defaultRateLimitConfig,
-      },
+      ratelimit: {
+        config: {
+          namespace: "exceeded",
+          ...defaultRateLimitConfig,
+        },
 
-      ratelimitIdentifierFn: (_req: MiddlewareRequest) => {
-        throw new Error("Error simulated by test");
+        ratelimitIdentifierFn: (_req: MiddlewareRequest) => {
+          throw new Error("Error simulated by test");
+        },
       },
     };
     const unkeyMiddleware = withUnkey(options);
@@ -174,18 +186,20 @@ describe("withUnkey", () => {
     // here the the identifier function will throw an error
     // which will cause the default error response to be returned
     const options: withUnkeyOptions = {
-      ratelimitConfig: {
-        namespace: "exceeded",
-        ...defaultRateLimitConfig,
-      },
+      ratelimit: {
+        config: {
+          namespace: "exceeded",
+          ...defaultRateLimitConfig,
+        },
 
-      ratelimitIdentifierFn: (_req: MiddlewareRequest) => {
-        throw new Error("Error simulated by test");
-      },
-      ratelimitErrorResponseFn: (_req: MiddlewareRequest) => {
-        return new MiddlewareResponse("Custom Error message when rate limiting", {
-          status: 500,
-        });
+        ratelimitIdentifierFn: (_req: MiddlewareRequest) => {
+          throw new Error("Error simulated by test");
+        },
+        ratelimitErrorResponseFn: (_req: MiddlewareRequest) => {
+          return new MiddlewareResponse("Custom Error message when rate limiting", {
+            status: 500,
+          });
+        },
       },
     };
     const unkeyMiddleware = withUnkey(options);
@@ -202,18 +216,20 @@ describe("withUnkey", () => {
     // here the the identifier function will throw an error
     // which will cause a custom error response to be returned
     const options: withUnkeyOptions = {
-      ratelimitConfig: {
-        namespace: "exceeded",
-        ...defaultRateLimitConfig,
-      },
+      ratelimit: {
+        config: {
+          namespace: "exceeded",
+          ...defaultRateLimitConfig,
+        },
 
-      ratelimitIdentifierFn: (_req: MiddlewareRequest) => {
-        throw new Error("Error simulated by test");
-      },
-      ratelimitErrorResponseFn: (_req: MiddlewareRequest) => {
-        return new MiddlewareResponse("Not implemented", {
-          status: 501,
-        });
+        ratelimitIdentifierFn: (_req: MiddlewareRequest) => {
+          throw new Error("Error simulated by test");
+        },
+        ratelimitErrorResponseFn: (_req: MiddlewareRequest) => {
+          return new MiddlewareResponse("Not implemented", {
+            status: 501,
+          });
+        },
       },
     };
     const unkeyMiddleware = withUnkey(options);
@@ -228,22 +244,24 @@ describe("withUnkey", () => {
 
   it("should use a RedwoodJS-compatible logger", async () => {
     const options: withUnkeyOptions = {
-      ratelimitConfig: {
-        namespace: "exceeded",
-        ...defaultRateLimitConfig,
+      ratelimit: {
+        config: {
+          namespace: "exceeded",
+          ...defaultRateLimitConfig,
+        },
+
+        ratelimitIdentifierFn: (_req: MiddlewareRequest) => {
+          throw new Error("Error simulated by test");
+        },
+        ratelimitErrorResponseFn: (_req: MiddlewareRequest) => {
+          return new MiddlewareResponse("Not implemented", {
+            status: 501,
+          });
+        },
       },
       logger: createLogger({
         options: { level: "error", enabled: true },
       }),
-
-      ratelimitIdentifierFn: (_req: MiddlewareRequest) => {
-        throw new Error("Error simulated by test");
-      },
-      ratelimitErrorResponseFn: (_req: MiddlewareRequest) => {
-        return new MiddlewareResponse("Not implemented", {
-          status: 501,
-        });
-      },
     };
     const unkeyMiddleware = withUnkey(options);
     const request = new Request("http://localhost:8910/api/user");
